Prevent other tabs

+++ scripts/update_version.py
@@ -1,3 +1,4 @@
+from datetime import datetime
 import getopt
 import json
 import os
@@ -37,31 +38,54 @@ def get_next_version(current_version: str, part_to_update: str):
     return f"{major_version}.{minor_version}.{bugfix_version}"
 
 
-def write_new_version(file_path, new_version):
+def write_new_version(file_path: str, new_version: str, dry_run: bool = False):
     with open(file_path, 'r') as f:
         data = json.load(f)
 
     data['version'] = f"{new_version}"
 
+    if dry_run:
+        return
+
     with open(file_path, 'w') as f:
-        json.dump(data, f, indent=4)
+        json.dump(data, f, indent=2)
+
+
+def write_changelog_update(file_path: str, new_version: str, dry_run: bool = False):
+    with open(file_path, 'r') as f:
+        data = f.read()
+
+    if data.find(new_version) < 0:
+        now = datetime.now()
+        date = now.strftime('%Y-%m-%d')
+
+        data = f"### {new_version} - {date} - <title>\n\n- <insert changes>\n\n{data}"
+        print(data)
+
+        if dry_run:
+            return
+
+        with open(file_path, 'w') as f:
+            f.write(data)
 
 
 def main():
     argument_list = sys.argv[1:]
-    options = 'p:'
-    long_options = "part="
+    options = 'dp:'
+    long_options = ["dry-run", "part="]
     # Parsing argument
     arguments, values = getopt.getopt(argument_list, options, long_options)
     part = 'bugfix'
+    dry_run = False
     # checking each argument
     for current_arg, current_value in arguments:
-
         if current_arg in ("-p", "--part"):
             if current_value in ('major', 'minor', 'bugfix'):
                 part = current_value
             else:
                 print("No part provided. Updating the bugfix version by default")
+        if current_arg in ('-d', '--dry-run'):
+            dry_run = True
 
     git_root = get_git_root(os.getcwd())
     files = [git_root + '/public/manifest.chrome.json', git_root +
@@ -86,7 +110,9 @@ def main():
     print(next_version)
 
     for file in files:
-        write_new_version(file, next_version)
+        write_new_version(file, next_version, dry_run)
+
+    write_changelog_update(git_root+"/CHANGELOG.md", next_version, dry_run)
 
 
 if __name__ == "__main__":
